pyproject.toml
[project]
name = "email-services"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "imapclient>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "secure-smtplib>=0.1.1",
    "uvicorn>=0.35.0",
    "slowapi>=0.1.9",
    "pydantic[email]>=2.0.0",
    "cryptography>=41.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=23.0.0",
    "prometheus-client>=0.19.0",
    "psycopg2-binary>=2.9.10",
]
